pyult ver0.2.0
1. A new argument "size" to pyult.image.to_square, which specifies
lengths of sides of the result image.

2. A new function "pyult.image.time_to_frame", which converts (a list
of) time to its (their) corresponding frame(s).

3. A new method "pyult.recording.filter_by_segments", which extracts
only the frames that correspond to the segments specified by the
method's argument "segments".

4. pyult.recording.imgs_to_df and pyult.recording.square_imgs are
updated to have a new argument "inplace". With the argument being FALSE,
a recording object is returned.

5. pyult.Session.load_recording is updated and now returns a Recording
object with a new attribute "rec.ID". "rec.ID" is given through the
argument "basename_or_index" of pyult.Session.load_recording.

6. Bug fix, when xxx.words and xxx.phoneswithQ exist and read to
attributes of the Recording class.